Cleanup and consolidate initialization of the RNGs (#1560)
* Cleaned up the initialization of the random seed in LBANN library and
associated front ends.  Since the seed can be set by the user for each
trainer, the random seed should be initialized by the construction of
the trainer.  Remove the multiple RNG initializations that occur in
the base.cpp initialize function and multiple times in the
lbann_library construct trainer function.  Now the random seeds are
set by the user (possibly) and then have the trainer and global IDs
mixed in appropriately.  Also fixed a bug where the trainer rank was
never actually properly mixed in with the hash_combine function.  This
also adds the reporting of the user specified parameters to the
contruct_trainer function and properly reports the random seed that is
used.

* Fixed the tests to not require initializing the RNG if not needed.

* Added explicit flags and guards to ensure that the RNGs are not used
before they are seeded and initialized.

* Fixed a typo.

* Changed the cosntruction of the trainer and data coordinator classes
to receive the data readers during setup.  This fixes an ordering
dependency between the construction and setup of the data readers with
the initialization of the RNGs.  Now the trainer and RNGs can be
constructed and initalized before the data readers are constructed and
the data has it's preliminary shuffle.  This also fixes a bug where
the data readers will shuffle data with the incorrect (early) random
seed.

* Fixed the construction of the resnet and alexnet integration tests so
that the validation set is properly marked as the test set.

* Revised the performance numbers for the integrated test since LBANN_DETERMINISTIC is on

* Fixed a test file

* Put a guard around the generation of the Elemental seeding in the
initialization function.  Preserved code that uses MPI or LBANN comm
objects to mix rank info into the seed to ensure that Hydrogen
has good RNGs.  Note that under some configurations LBANN
will not do this, so it is good to ensure that Hydrogen is
well seeded.

* Updated the sequential catch tests to explicitly seed the LBANN random
number generators.

* Moved the global random number generators into a separate file that is
independent from the checkpoint and restore logic of the RNGs and the
generation of numeric distributions.  This is a step forward to
ensuring that each file / class only needs to include what it uses.
Updating classes that only access the RNGs to include the new RNG
class.

* Minor cleanup

* Capture the root random seed as well as the random seed after it is
modified by the trainer ID, etc and include it in the checkpoint and
restart.  Add mechanisms to gather every ranks' random seeds and allow
the world master to display the random seeds when printing the run's
parameters.  Added a function in LBANN library to define standard
options using the new parser class.

* Updated the front ends to include the new argument parser and standard
options.

* Reverted check to look for test top-5 accuracy not validation

* Fixed the expected time for Lassen runs of Lenet now that it is running with LBANN_DETERMINISTIC