Pair RNG with I/O thread work (#1575)
* Fixed how the data reader dispatches blocks of work to the
rest of the I/O thread pool and identifies blocks.  The I/O thread executing the
fetch_data_in_background function is executed by a thread in the
thread pool.  Therefore it has to dispatch work to other threads in
the thread pool and do some work locally. The fetch_data_block function 
now properly renames the
parameter thread_id to block_offset to indicate that it marks which
block of work is being executed.  The fetch_data_block now sets a
thread local variable to index into the new vector of I/O random
number generators.  The I/O random number generators are no long
thread local variables, but there is a global vector of them that is
paired with each block of I/O and transformation work.  The number of
I/O RNGs is setup during initialization to have one per thread in the
I/O pool.

* Apply suggestions from code review

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Improved how the I/O RNGs are packed into a vector to reduce
thread contention and improve performance.  Updated initialization of
I/O RNGs to occur at init time.  

* In the I/O RNG structure added an atomic field for the process ID 
and checking to ensure that only a single thread access the I/O RNGs 
at a time. The active thread ID is atomically set
and if another thread was using the RNG, LBANN will throw an exception.

* Switched to using a RNG reference class that provides a lock while in
scope and releases the lock after destruction.  This simplifies the
use of the I/O RNG class so that acquisition of the reference struct
atomically sets the thread ID.
* Properly reset the thread ID for the I/O RNG when done with the basic
block.

* Fixed the checkpoint and restart logic for the new vector of I/O RNGs.

* Include the hyperthreads in the core mask when launched on pascal.

* Replaced the old num_io_threads command line argument with a new
version based on the arg parser class so that we can override it with
an environment variable.

* Added code to setup and lock the I/O RNGs for the transform class tests

* Added the necessary includes and scoping for the new function to
access the I/O RNGs.

* Added a flag to allow the driver to create as many I/O threads as
there are free cores, rather than the number of cores in the CPUSET.

* Fixed a compile time error in debug mode introduced by the mixed
precision weights PR.

* Improved error reporting.

* Cleaned up naming per review comments

* Replaced the use of the pthread API to bind threads to CPU cores with
the hwloc interface.  Added code for each rank to identify cores local
to the active GPU and pin the I/O threads to those cores.  For
CPU-only systems, grab all available cores.  All threads are bound to
the entire processing unit (PU) pool available to reach rank.

* Adjusted training times for the integrated tests to account for the thread-safe RNGs.

* Adjusting times on catalyst

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>