Optimize performance of GRU layer (#1641)
* Add wrapper classes for CUDA graph API objects

* Progress toward using cuDNN 8 RNN API

* CUDA graph support in GRU layer forward prop

* GRU layer uses cuDNN 8 RNN API

* Minor tweaks

* Support multiple layers in GRU layer forward prop

* Support stacked GRU in GRU layer

* Update ATOM VAE model to use stacked GRU layer

* Reuse CUDA graph in GRU layer

* Minor tweaks to GRU layer

* Fix bug in GRU layer backprop where graph was recomputed every step

* Clean up

* Fix compile error on systems with old CUDA or cuDNN

* Update release notes

* Added a cache for GRU cuda graphs.  Both forward and backward prop
have unordered maps that contain a computed graph for each mini-batch
size.  Using the mini-batch as the key ensures that if a phase has new
buffers but the same mini-batch size, it will overwrite the existing
graph cache, avoiding memory leaks.

* Fix bug where GRU layer recomputed graph every backprop step

* Clean up

* Pad to minimum batch size in GRU layer

cuDNN has bug where kernels fail if mini-batch size is small.

* Refined the fix to the bug in the GRU optimization that bug seems to
be triggered by switching mini-batch sizes when the last mini-batch is
too small.  This means that we need a lower bound, which is
emperically tested to be 128 for WAE.  However, if the initial
mini-batch size is less than 128 and it isn't changed, things seem to
be okay.  So set the threshold to be the smaller of the initial
mini-batch size or 128.

Co-authored-by: Brian C. Van Essen <vanessen1@llnl.gov>