Move I/O buffers and control to data coordinator (#1538)
* Created a buffered data coordinator child class to hold data type
specific I/O buffers in the data coordinator.  This moves the
functionality for holding the data out of the input layer.

* Revamped the creation of the data coordinator so that it is
instantiated independently and passed into the trainer class.  Added
support to select the data type required for the buffered data
coordinator.  Removed the trainer pointer from the data_coordinator
and provided an observer pointer to the IO thread pool.

* Removed old forward declaration

* Fix the temporary state to pull the labels into the I/O buffer.  This
needs to be generalized so that labels and response are fetched.

* Removed unused and masking checkpoint calls in the buffered data coordinator

* Moved the responsibility for fetching data in the foreground and
background into the data coordinator and out of the input layer.  The
execution context will now request that the data coordinator fetch
data prior to the forward pass and check if the epoch is complete and
kickoff any required background I/O before the backward pass.

* Moved the dump_mb_sample_indices to the data coordinator from the
input layer.

* Moved the code for accessing the size of the data set, current, and
global mini-batch size out of the input layer and into the data
coordinator.

* Removed most of the access functions for getting to the data set
statistics objects from the input layer.  These functions are now in
the data coordinator.

* Switched access to check the number of samples trained and tested from
the input layer into the data coordinator.

* Moved access to the data readers entirely to the data coordinator and
out of the input layer.

* Moved reaping of the I/O threads to the destruction of the data
coordinator rather than the input layer.

* Updated the check_gradients callback to properly fetch data.

* Consolidated the io_layer, generic_input_layer, and input_layer into
the input_layer class.

* Moved the input layer up in the file hierarchy to the io directory.

* Moved the fetch_to_local_matrix from the partitioned_io_buffer class
into the buffered_data_coordinator class.  This is the first step in
collapsing the io_buffer class hierarchy.  The I/O data_buffer objects
are now contained in a map with a new enum class called
input_data_type that is used to identify SAMPLES, LABELS, and
RESPONSES data types in a data reader.  The data readers will now
describe which of those input types are supported for each instance of
a data reader.  The buffered_data_coordinator will now fetch all
available data for each sample, namely the sample data, any labels,
and any response values.  This then allows input layers to selectively
access any of these fields.

* Removed deprecated field from prototext

* Created a top level function in the data reader to return all of the
data for each sample in a single fetch call.  The
buffered_data_coordinator uses this interface now and passes a set of
local matrices into the data reader for storage.

* Transferred the functionality of the partitioned I/O buffer into the
buffered data coordinator and removed both the partitioned and generic
I/O buffer classes.

* Cleaning up deprecated includes.  Only include data dimensions if the
data reader supports it.

* Removed unnecessary include for atomics.

* Added the io_data_buffer that takes the data buffer clas soutof the
partitioned I/O buffer class.

* Fixed the setup and construction of the buffered data coordinator to
match the new approach in develop for the data coordinator.  Added
const and non-const access methods for several of the buffered data
coordinators io buffers.

* The model no longer holds the data readers so the data coordinator is
queried to see if the execution mode is valid.

* Updated the summarize images callback to use the data coordinator
rather than the input layer for getting information about the
mini-batch.

* Changed the non-const path for getting the active buffer to use a
direct implementation rather than casting and calling into the
const-path.

* Removed Cereal atomic header.

* Fixed the return type of the data coordinator responding to the number
of iterations per epoch to match the unsigned type of the execution
context.

* Fixed up merge conflict.

* Changing the behavior of the input layer so that it defaults to a data
reader target mode of NA rather than CLASSIFICATION. This means that
the input layer will expect to only have one output, unless the mode
is set to CLASSIFICATION, REGRESSION, RECONSTRUCTION, or
LABEL_RECONSTRUCTION (all of which have two outputs).  The prototext
default is also set to match this behavior.  For all prototext and
Python Front-End applications that require labels, it is now necessary
to explicitly set the mode to classification using the
target_mode='classification' flag (classification was the old
default).  All other models should be fine, since they would have
previously explicitly set this.  This new behavior is in line with the
expectation of the python data reader, which typically ingests all
information as a packed sample.

* Added the target_mode: "classification" flag to the input layers.

* Moved the serialize_io flag from the model class to the trainer class.
This change appears in the lbann library, prototext and python front
end.

* Trainer now owns serialize_io

* Added guards in data reader fetch function to ensure that each buffer
for SAMPLES, LABELS, and RESPONSES are valid.

* Added guards to the buffered_data_coordinator to only request labels
or responses if the data reader has them.  Additionally, fixed a bug
on the guard in fetch_to_local_matrix to make sure that the local
matrix has space, not just the global matrix.

* Updated the data coordinator to use execution_mode iterators to scan
over all available data readers.  Changed the get_data_reader function
to allow any execution mode to return a valid data reader or nullptr,
rather than throwing an exception for phases without valid data readers.

* Updated input layers to use flage target_mode = 'classification' as
appropriate.

* Code cleanup.

* Fixed a bug in the buffered_data_coordinator's
get_sample_indices_per_mb call where the const_cast was creating a
copy of the class.  Fixed the dump_mb_sample_indices callback couldn't
create the directory and the output files were overwritten by each rank.

* Updated Conduit to v0.5.1 to address older bugs.

* Moved the serialize_io function into the trainer construction.
Updated construction of the I/O thread pool to enable this flag.

* Added a note about the reason for putting the background data fetch in
the epoch_complete function.

* Apply suggestions from code review

Co-authored-by: Tim Moon <moon13@llnl.gov>

Co-authored-by: Tim Moon <moon13@llnl.gov>