Improved spack external definitions (#1699)
* Added feature to build required spack packages for building LBANN
documentation.  Specifically, breathe, sphinx, and sphinx-rtd-theme.
Also added a feature to use a generic spack externals file that is
selected via spack platform and then specialized by the full spack
architecture.  If there is a more specific external file, that will be
preferred over a generic external.

* Updated the install_lbann.sh script to look for a set of external
definitions using a combination of the spack architecture platform and
a generic target architecture from uname -m.  This allows common
definitions of externals for generic target architectures.  Renamed
externals for x86_64 and ppc64le.

* Added a spack specific way to find the target architecture family.

* Consolidated all of the x86_64 externals into a common file.

* Updated the build environment for OS X to Catalina.

* Added support to define a CUDA compute architecture for the main LBANN
build and the dependencies.  Also updated pascal to use CUDA 11.1 and
cuDNN 8.0.4.  Cleaned up and refactored the code used to identify the
compute center so that both the install_lbann.sh and
build_lbann_from_source.sh use the same code snippets.

* Updated the minimum spack version to v0.16.0