Feature modular trainer split (#1719)
* Split the task of splitting the MPI communicator out of the function
that constructs the trainer.  This allows upcoming changes to the LTFB
algorithm to allow it to use a separate prototext per trainer.  This
required removing the procs_per_trainer field from the trainer
prototext and it is now only a command line argument.

* Updated PFE to pass the procs_per_trainer as a command line argument,
rather than placing it in the trainer prototext.

* Fixed python typos and removed old help message about
procs_per_trainer.

* Added the new arg_parser to the dedicated help front end.

* Removed dead code

* Fixed up how the trainer is split and removed dead code.

* Updated the construct_trainer and print_parameters functions to get
the procs_per_trainer from the comm object.