defer layer graph serialization until the end of model serialization
The intent is to limit the stack depth. As Cereal uses a depth-first
algorithm, the entire model is essentially serialized by just
serializing any single layer -- this should capture the whole layer
DAG as well as all weights (assuming every layer is in the dag, a
model contains only one dag, and every weights is associated with at
least one layer). This is potentially a serious problem -- I observed
serializing Lenet reached 959 frames and it wasn't done yet -- and
Lenet is a tiny network.

Essentially, the decision to defer these means that the actual
serialization of the weights and layers happens via the vector of
owning pointers and then the deferred components are the "viewing"
parent/child information in each layer. I'm not sure (a) if there is a
cost of deferring serialization and (b) how severe that cost could
be. If we run into problems with larger models, I have a few other
ideas for how to work around the graph issue.

Currently the model calls `serializeDeferments()`, but it's not
difficult to imagine scenarios in which multiple models might be
stored in the same archive. Since `serializeDeferments()` is to only
be called on an archive object once during its lifetime, I think I
will eventually move that out of the model and push that onto the
caller (and suitably document that decision).