Move model checkpointing from end of training step to beginning (#1733)
* Move model checkpointing from end of training step to beginning

* Changed all of the checkpoint logic to checkpoint at the start of
epochs and changed the naming scheme to use the callback phase (via
the visitor hook) in the name rather than the current execution
context.  Using the callback hook identifier rather than the current
execution mode provides more descriptive names for the resulting
checkpoint files.  Added an enum type of visitor hooks that will
eventually be used to allow callbacks or other visitors to operate at
user defined hook points.

* Removed commented includes

* Have the training algorithm reset the execution context back to
training before the end of training callbacks are called.  This ensure
that any callback sees the correct execution context.

* Updated the visitor_hook enum data type and the callback to use a
generic execution_mode_begin and _end pattern so that the visitor
pattern is not specific to a given execution mode, but can mixed with
the execution_mode.  The new hooks are prefixed with the
execution_mode string and are combined with the actual execution_mode
to for phase specific strings, etc.

* Default value for mode in read_latest

Co-authored-by: Brian C. Van Essen <vanessen1@llnl.gov>