Feature reworked install process (#1732)
* Updating the developer approach for building LBANN to use newer Spack
capabilities spack dev-build and spack build-env.  This will allow us
to more easily rely on the spack recipes for LBANN and associated
software for configuring local CMake, etc.

* Cleaning up the build_lbann.sh script.  Improving the help message and
adding modules for other LC systems.

* Fixed the code to find and uninstall older versions of lbann with the
same label on the architecture.

* Removed debugging

* Fixed the x86_64 modules so that they also pull StdEnv from the module
path so that the correct CMake is found.

* Added an option for stopping after spack spec is called.

* Added modules for Corona.

* Added an guard for when no pre-specified modules were found.

* Added spack commands to find the compilers and put them into the
environment.  Also added flags for Haswell architecture.

* Cleanup of the script and add modules for Catalyst Ivybridge system.

* Added the ability to stop after installing the dependencies.  Added
options to use local repostories for Hydrogen, DiHydrogen, and Aluminum.

* Updated the documents to use the new build_lbann.sh script.

* Added default modules for OLCF and NERSC.

* Bumped up the version of gcc used on Summit

* Rearranged build script so that the help option works without Spack.
Throws a good error message if SPACK_ROOT isn't set and removed
superfluous parens.

* Added guard to check for the presence of the environmennt before
activating it.

* Added a new center specific function to patch in a small number of
externals that are not found by the spack external find command. Fixed
a bug in the build script where the MPI field was set after the Spack
command was generated.  Updated the new build recipe for CoriGPU to
work with the latest available modules.

* Added support for local unit tests.

* Fixed the spack env grep command to not find substrings that match the
environment name.  Also added the extra commands like --test root to
both the install and dev-build commands.

* Updated the build script to output the resulting command to drop back
into the build environment so that they can rebuild easily.

* Added hints on how to rebuild LBANN to the output of the script.

* Added support to build lbann source in a temp directory to speedup
compile and link times.

* Updated the documentation to reflect the new build scripts.

* Added a quick sed script to add gfortran supporting compiler to any
clang compilers found.

* Moved the specification of the GPU_ARCH_VARIANT into a GPU_VARIANT
guarded block.

* Fixed guard on symlink removal

* Updated the user installation instructions.

* Fixed the build script so that it doesn't uninstall the dev-build of
LBANN unless the dependencies are being reinstalled.  Also added help
text for using the module command to load LBANN.

* Fixed where the removal of old installed instances of LBANN are in
Spack by doing it after an old environment may be removed and before
the environment is activated.  Additionally, it has to be removed even
when dependencies are not installed, it is the spack dev-build command
that installs it into Spack.  Finally, added a helpful hint at the end
to show how to load the module file for the installed LBANN instance.

* Added the use of the spack add and develop commands for the LBANN
package so that the environment properly picks it up.  Note that the
dev-build command does not do this, it is not environment aware.

* Moved the calculation of the hash for LBANN to after possible addition
of develop targets for Hydrogen, DiHydrogen, or Aluminum.

* Switched the order of when the dev path for LBANN is included to
ensure that it doesn't get paired with subsequent dependencies.

* When using mvapich2 on LC systems, it needs to be paired with
hwloc@1.11.13, since it was build against that.  Also only the Power
systems have a problem when building their own rdma-core library and
have to use an external.

* Fixed the bamboo tests so that they do not take hard-coded executable
paths to find the LBANN executable.  Removed code to try and find
versions of LBANN explicitly compiled with gcc7, intel19, and clang.
Note that the clang and intel compiler tests were not in use and also
led to a large number of skipped tests.  Fixed a bug in the unit tests
invocations that led to redundant instances of the test functions for
each test.

* Removed option for passing explicit executable into the local tests.
Based on the new test environment the executable and PYTHONPATH
should be in your environment.

* Updated the run script to load the module built by the compiler part
of the test.  Updated the compiler test script to only build with a
single compiler when using the cmake build script.

* Updated the build script to save the log file in the build directory
after the build is complete.

* Fixed typo

* Updated the Spack environment to use a module naming scheme that
places the version in a hierarchical manner rather than appending to
the package name.  This provides better package collision resolution
within Spack and with regular lmod.  Removed support for the old spack
build system, including the preconfigured center-specific spack
environments.  This functionality is achieved by both spack externals
find and a bit of hard-coding certain packages for each center.

* rename modulefile to provide proper lmod version collision behavior

* Cleaned up the instructions on how to load the module

* Updated the docs to indicate the ability to load with module commands

* Fixed architecture name for Corona Zen system

* Clarified help message. Changed test for existing link to check for
a file name of any type.

* Removed the feature for the LBANN module to autoload other modules.
This caused weird bugs where dependent modules were unable to be
loaded and caused LBANN to fail loading.  Changed the python call to
explicitly look for a python3.x executable.

* Simplified the build script to rely more heavily on the "version"
label flag for naming the resulting environment and build directory
link.  There are no longer explicit options to set the spack
environment name or the build directory.  The lbann version label now
has the spack architecture appended to it to make it easier to track
on multi-architecture instances of Spack.

Changed how existing versions of spack are found to avoid using the
spack architecture because of a corner case where a compiler forced a
"simpler" architecture build because it couldn't support more advanced
features.  This cause the script to inadvertently not be able to find
previous installations,

* Added helpful hints for reparing Spack modules.

* Added RIKEN and the FX700 system as a center and target platform

* Fixed a bug in the set_center_specific_externals function to also take
the spack architecture, not just the target to ensure that the
externals have the proper architecture field.

Updated the modules used for CoriGPU at NERSC.

* Cleaned up the documentation to reflect changes to the way that
version labels are used to generate the environment and build links.

* Added a tip about version labels.

* Tweaked documentation

Co-authored-by: Thomas R. Benson <benson31@llnl.gov>