Updated the build script for ROCm on Corona (#1792)
* Updated the build script for ROCm on Corona.  Added external paths for
the HIP/ROCm tools.

* Updated the ROCm build on corona to use clang 11.0.0 by default.
Added a PERL hack to add in a gcc-toolchain flag for Clang C and CXX
compilers to enusre that on LC systems use a modern version.  Also
updated script to use ROCm-4.0.0.

* Added externals for the ROCm / HIP software stack.  Updated the
ordering of adding manual externals to proceed allowing spack to find
external compilers or packages.  This is necessary to accommodate the
formating of the yaml file.  Separated the Clang compiler cleanup to
a separate function.  Added a perl fixup for Clang to use a modern gcc
library on LC systems.

* Fixed arch target

* Cleanup modules for Corona

* Make BOOST an optional dependency to avoid build issues on Corona

* Fix how DiHydrogen is handled in the build script. Added reasonable
default dependencies for Power systems.

* Renamed the helper function to indicate that it sets Center specific
dependencies, not just MPI.  Also added the GPU variant to DiHydrogen
to deal with Spack concretization issues.