make statistics mutable, finish constification
The fundamental hypothesis here is that, logically, a communication
does not mutate a communicator. Even bitwise, the only modification to
the lbann_comm with a communication is the accumulation of
statistics. This is the textbook use-case for "mutable" parameters --
the logical state about the communicator (its group and context) have
not been modified in anyway and the bitwise changes to the object do
not affect the invariants of its class.