Cleanup of some of the comm guts; fix const correctness issues. (#1807)
* add noexcept decorations throughout lbann_comm iterface

* fix a plethora of const/noexcept issues in lbann_comm

* rename data members to comply with data member naming policy (leading "m_")

* make statistics mutable, finish constification

The fundamental hypothesis here is that, logically, a communication
does not mutate a communicator. Even bitwise, the only modification to
the lbann_comm with a communication is the accumulation of
statistics. This is the textbook use-case for "mutable" parameters --
the logical state about the communicator (its group and context) have
not been modified in anyway and the bitwise changes to the object do
not affect the invariants of its class.

* fix missed member variable renames

* fix incorrect const decorations