Updated the build script to use the new environment aware feature of (#1824)
Spack that allows the spack build-env lbann -- bash command to find
the proper version of LBANN that is in the active environment.  This
simplifies the user interface so that the develop does not have to
remember the full set of Spack variants.  As a result the build script
flag --build-env-only is deprecated.  Additionally, cleaned up the
implicit dependencies on Hydrogen and DiHydrogen to properly work when
Aluminum is disabled.