Try to fix testing.
NOTE: training_algorithm::get_type() is not entirely correct
here. It could cause problems if multiple instances of the same
callback are checkpointed at the same time, but currently we
don't do this. The issue with "get_name()" is that we might not
know the name of the training algorithm on restore (since the
training algos themselves are not checkpointed).