Reformatting, documentation, and slight API tweaks (#1827)
* formatting for training_algorithm and sgd_training_algorithm

also includes some updated documentation for training algorithms, as
well as a slight API modification ("get_name" is now "get_type", and
names are user-specified).

* cleanup some issues with the training_algorithm use in trainer

* termination_criteria are now properly polymorphic

* one more API tweak

* Revert "cleanup some issues with the training_algorithm use in trainer"

This reverts commit 8afcc998d0ee3e7d98e0961b9bac2f4d32a789ac.

* Patch up the previous commit

* more docs

* formatting for execution contexts

* fix some things that got lost in the git acrobatics

* Try to fix testing.

NOTE: training_algorithm::get_type() is not entirely correct
here. It could cause problems if multiple instances of the same
callback are checkpointed at the same time, but currently we
don't do this. The issue with "get_name()" is that we might not
know the name of the training algorithm on restore (since the
training algos themselves are not checkpointed).