back out changing save/load function to return nothing
The return value of these functions is never considered. However,
unraveling this mess should be outside the scope of this PR. In a
future PR, we should consider the behavior we want to see and more
carefully treat the difference between "ok to fail" and "cannot
fail". Ideally we will use exceptions for clarity.