hack around a static initialization/destruction bug
In at least one scenario caught by our testing, the hydrogen memory
pools were being destroyed before the trainer object. This caused
problems -- an exception in the best case (which triggered
std::terminate since it was in a destructor) and a hang in the worst
case (trying to lock a destroyed mutex). This should fix *that* issue,
but we should clean this up and be more careful going forward.