Cleanup execution context (#1849)
* cleanup trainer API and documentation; push related changes through codebase

* add get_trainer() method to access global trainer; apply to frontends

* finish cleaning up the execution context

I still want to pare it down more, but at least the base class is
default-constructible now.

* remove extraneous header inclusion

* fix with destroying a trainer

The data coordinator must be destroyed before the thread pool. We
should manage this dependent lifetime differently (IMO, the
data_coordinator should probably manage the I/O threadpool).

* back out changing save/load function to return nothing

The return value of these functions is never considered. However,
unraveling this mess should be outside the scope of this PR. In a
future PR, we should consider the behavior we want to see and more
carefully treat the difference between "ok to fail" and "cannot
fail". Ideally we will use exceptions for clarity.

* hack around a static initialization/destruction bug

In at least one scenario caught by our testing, the hydrogen memory
pools were being destroyed before the trainer object. This caused
problems -- an exception in the best case (which triggered
std::terminate since it was in a destructor) and a hang in the worst
case (trying to lock a destroyed mutex). This should fix *that* issue,
but we should clean this up and be more careful going forward.