Build script uses clingo (#1846)
* Updated the documentation to tell the user to install and use the new
clingo concretizer for Spack.  This allows us to simplify the
specification of build dependencies. Updated the build script to use
the sequence of commands inside of an environment:

spack add lbann
spack develop lbann <path>
spack install

rather than using the command spack dev-build because it isn't
actually environment aware.

Removed additional variant specifications for Hydrogen, DiHydrogen,
and Aluminum that are no longer needed with the new concretizer.

* Removed deprecated spack dev-build command

* remove references to dihydrogen as an explicit variant

* Update docs/building_lbann.rst

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Updated the failure message to report the name of the environment and
build directory when the script catches an error.

* Updated the build script to automatically update your Spack repository
to use clingo.  Modified the documents to reflect this.

* Added a help message for when spack solve fails due to a lack of a
modern compiler.

* simple typesetting changes

* Change the build script so that it passes either the gold or lld
linkers to Spack rather than having Spack include it for the LBANN
package.  This improves the LBANN recipe and avoids expectation that
the base system has either of these linkers available.

Co-authored-by: Thomas R. Benson <benson31@llnl.gov>
Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>