Make build_lbann.sh script extensible (#1866)
* Added a new --extras or -e flag to allow users to include a file of
Spack packages that should be included in the environment.  Added an
option to only configure and to print help directions when only
configuring.  Also cleaned up some of the other flags and their help
message.

* Added a option (-p) to specify multiple additional packages on the
command line that get added to the environment and concretized with
LBANN.

* Added a flag to use the build script to install as a "user".  This
means that the code is fetched from GitHub and doesn't use a local
repository.  It does still setup the environment and standard libaries
and modules on known systems.  By default it will build with the
latest named versions of LBANN, Hydrogen, DiHydrogen, and Aluminum.
It takes an optional argument to specify a version label.  For user
builds the version label is independent of the spack environment name.

* Added run instructions to the function exit_with_instructions

* Switched the command to find the Spack hash of the LBANN package to
use spack concretize rather than spack solve to account for the
presence of other top level packages.

* Added guards to the build script so that certain file system tasks
aren't performed during a user build.

* Added a note to the documentation about how a user could download and
use the LBANN build scripts to configure the default external packages.

* Added a new user facing build script that can be executed directly
from a sequence of curl'ing the script into bash.  (Similar to how
Homebrew installs.)  Added a default set of variants to enable based
on the Center and system being targeted.

* Updating the script to use a configurable path for finding the Git
repo.  Added a temporary path to the working branch.

* Look for the customize script locally

* Added the parallel build command for users

* Set the user build to use stable dependencies for Hydrogen, Aluminumm, and DiHydrogen.

* Corrected how the MIRRORS are tested when adding multiple.  Creating
the Spack build cache now takes its own path and is tested
independently.  Added a preliminary spack concretize to make sure that
there is a valid configuration before finishing configuration and to
make sure that any error codes can be caught.

* Enabled multiple Spack mirrors in the build script.  Added a default
clang@amd compiler for Corona.  Updated the CENTER_FLAGS to turn on
either the gold of lld variants in the updated Spack LBANN package so
that the ldflags are not propagated to all dependents.  Fixed the sed
script for adding Fortran support to clang compilers to find the new
null string that Spack switched to using.

* Fixed a bug where there MIRRORS weren't being properly checked

* Ensure that spack mirrors are unique

* Keep the config padding if there are mirrors available

* Fixed the build script to use spack find for the spec hash, rather than another concretize command.

* Added an ATOM specific user build script

* Updated the spack find command to find the LBANN spec's hash using a
new flag to get the hash of a concretized, but not yet installed
version of LBANN.

* Changed wget command to move existing files to backup files

* Fix wgets backup naming scheme

* Switched over to using curl rather than wget for pulling down files
from GitHub.  Removed the concretize-only option since the
build_lbann.sh script already has to concretize at least once to enure
that the environment is valid and the error code of that spack
concretize command has to be tested.

* Add any GPU architecture variants to PyTorch if it is included.

* Create a set of GPU variants to pass to any dependent packages

* Apply suggestions from code review

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Fixed RST

* Don't forward the amdgpu_target to PyTorch external package

* Updated the release notes and set curl paths back to the main repository

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>