Fixed a bug in the close_if_done_samples_file_handle function where the test operation looking to see if a file handle was valid prior to closing it was incorrectly inverted.
Renamed the close_if_done_samples_file_handle function to
close_samples_file_handle function and added a flag to allow the file
to remain open if it the file is predicted to be used soon, if the
flag is false the file will be closed regardless of potential future
uses.

Removed the pr_open_fd boolean flag that was used in the
manage_open_file_handles since it was confusing and unnecessary.
Removed the associated guard using the flag that skipped updating the
file_access_queue.

Provided an enum for accessing the file_id_stats_t tuple.

Updated the data_reader_sample_list shuffle_indices function to
compute the file access usage for each epoch after the shuffle.